server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "fastapi>=0.128.0",
    "inngest>=0.5.13",
    "llama-index-core>=0.14.12",
    "llama-index-readers-file>=0.5.6",
    "openai>=2.14.0",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "sqlmodel>=0.0.31",
    "streamlit>=1.52.2",
    "uvicorn>=0.40.0",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP"]

[tool.pyright]
typeCheckingMode = "basic"
reportMissingTypeStubs = "none"
